requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
DAWG-Python==0.7.2
docopt==0.6.2
fasttext==0.9.2
frozenlist==1.3.1
idna==3.4
joblib==1.2.0
multidict==6.0.2
nltk==3.8.1
numpy==1.23.4
pandas==1.5.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyaes==1.6.1
pyasn1==0.4.8
pybind11==2.10.3
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyTelegramBotAPI==4.7.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.4
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rsa==4.9
six==1.16.0
Telethon==1.25.4
tqdm==4.64.1
urllib3==1.26.12
yarl==1.8.1
